Update README.md
After login or signup, the page is an error, so refresh the page the chat page will appear. you can create a group and search for the user by clicking on the search button and ensure that the user you are searching for is logged in. You can message the user, but there is some problem with the real-time messaging, so after messaging the user, click on some other user and again click on the user you have messaged, and the message will appear.